requirements.txt
aiofiles
dropbox>=12.0.0
fastapi
jinja2
python-dotenv>=1.0.0
uvicorn
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fsspec
zstandard
s3fs
paramiko
